+++ src/pages/_app.page.tsx
@@ -2,7 +2,7 @@ import type { AppProps } from 'next/app'
 import React from 'react'
 import { Head } from 'components/common'
 import { WagmiConfig } from 'wagmi'
-import { appInfo, chains, wagmiClient } from 'utils/onboard'
+import { appInfo, chains, wagmiClient } from 'utils/chainClient'
 import { RainbowKitProvider } from '@rainbow-me/rainbowkit'
 
 import '../styles/antd.css'
